+++ TinkoffASDKUI/TinkoffASDKUI/Modules/SBPQr/Views/TableContentProvider/ISBPQrTableContentProvider.swift
@@ -1,0 +1,10 @@
+//
+//  ISBPQrTableContentProvider.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 18.04.2023.
+//
+
+import UIKit
+
+protocol ISBPQrTableContentProvider: ITableContentProvider where CellType == SBPQrCellType {}