+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * [MIC-8013] MainForm - Card payment logic
 * [MIC-8037] MainForm - Add analytics data
 * [MIC-8020] MainForm - Add localization
+* [MIC-8596] MainForm - Change SBP Button's background color
 * [MIC-7699] CardPayment - Add common functionality of payment by card
 * [MIC-8068] SBP - Opening from main form
 * [MIC-4650] CardsController & AddCardController for working with payment cards with own UI
@@ -43,6 +44,7 @@
 * [MIC-8050] Change add card icon
 * [MIC-8095] TinkoffPay - integrate to MainForm
 * [MIC-8096] TinkoffPay - integrate with independent button
+* [MIC-7776] PullableContainer refactoring
 
 ## [2.19.0] - 2023-04-13Z